Task 3.1.4
Выполнение задачи PP 3.1.4
