+++ src/main/java/ru/kata/spring/boot_security/demo/configs/SuccessUserHandler.java
@@ -16,10 +16,10 @@ public class SuccessUserHandler implements AuthenticationSuccessHandler {
     @Override
     public void onAuthenticationSuccess(HttpServletRequest httpServletRequest, HttpServletResponse httpServletResponse, Authentication authentication) throws IOException {
         Set<String> roles = AuthorityUtils.authorityListToSet(authentication.getAuthorities());
-        if (roles.contains("USER")) {
-            httpServletResponse.sendRedirect("/user");
-        } else if(roles.contains("ADMIN")) {
+        if (roles.contains("ADMIN")) {
             httpServletResponse.sendRedirect("/admin");
+        } else if(roles.contains("USER")) {
+            httpServletResponse.sendRedirect("/user");
         } else {
             httpServletResponse.sendRedirect("/");
         }
