+++ src/main/java/ru/kata/spring/boot_security/demo/controller/DataController.java
@@ -1,0 +1,43 @@
+package ru.kata.spring.boot_security.demo.controller;
+
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import ru.kata.spring.boot_security.demo.models.MyUser;
+import ru.kata.spring.boot_security.demo.repositories.RoleRepository;
+import ru.kata.spring.boot_security.demo.repositories.UserRepository;
+import ru.kata.spring.boot_security.demo.service.UserService;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/data")
+public class DataController {
+
+    private final UserService userService;
+    private final UserRepository userRepository;
+    private final RoleRepository roleRepository;
+
+    @Autowired
+    public DataController(UserService userService, UserRepository userRepository, RoleRepository roleRepository) {
+        this.userService = userService;
+        this.userRepository = userRepository;
+        this.roleRepository = roleRepository;
+    }
+
+    @GetMapping("/findOne")
+    public MyUser findOne(@RequestParam(value = "id", required = true) Long id) {
+        MyUser user = userRepository.findUserById(id);
+        user.setRolesInString();
+        return user;
+    }
+
+    @GetMapping("/findAll")
+    public List<MyUser> findAll() {
+        return userRepository.findAll();
+    }
+
+}
